src/main/java/itrillian/feedreader/rss/RssUtil.java
package itrillian.feedreader.rss;

import static itrillian.feedreader.xml.XmlAttributes.SCHEMA;

import java.io.File;
import java.io.IOException;

import javax.xml.parsers.DocumentBuilder;
import javax.xml.parsers.DocumentBuilderFactory;
import javax.xml.parsers.ParserConfigurationException;

import org.w3c.dom.Document;
import org.xml.sax.SAXException;

public class RssUtil {

	public static String getSchema(File file) throws ParserConfigurationException, SAXException, IOException {
		DocumentBuilderFactory dbFactory = DocumentBuilderFactory.newInstance();
		DocumentBuilder dBuilder = dbFactory.newDocumentBuilder();
		Document doc = dBuilder.parse(file);
		
		return doc.getDocumentElement().getAttribute(SCHEMA);
	}
}
